eleven_agent_acceuil/requirements.txt

python-dotenv
requests
PyYAML
